Add loading gif, connected loading to the store
