+++ src/index.js
@@ -5,7 +5,7 @@ import App from './containers/App/App';
 import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
 import { createStore, applyMiddleware } from 'redux';
-import rootReducer from './reducers';
+import rootReducer from './reducers/root-reducer';
 import { composeWithDevTools } from 'redux-devtools-extension'
 import thunk from 'redux-thunk';
